+++ react/SearchContext.js
@@ -51,7 +51,7 @@ const SearchContext = props => {
 
   const {
     params: { path: attributePath },
-    query: { query, map, order, operator, fuzzy, priceRange },
+    query: { _query, map, order, operator, fuzzy, priceRange },
   } = props;
 
   const url = useMemo(
@@ -66,9 +66,9 @@ const SearchContext = props => {
   );
 
   const initialVariables = {
-    query,
     operator,
     fuzzy,
+    query: _query,
     page: 1,
     store: account,
     attributePath: url,
@@ -115,7 +115,7 @@ const SearchContext = props => {
   };
 
   try {
-    if (!query) throw new Error("Empty search is not allowed");
+    if (!_query) throw new Error("Empty search is not allowed");
 
     return (
       <Query
@@ -132,7 +132,7 @@ const SearchContext = props => {
             error || !data
               ? VtexSearchResult.emptySearch()
               : new VtexSearchResult(
-                  query,
+                  _query,
                   1,
                   props.maxItemsPerPage,
                   order,
@@ -149,7 +149,7 @@ const SearchContext = props => {
             searchResult:
               error || !data
                 ? {
-                    query: props.params.query,
+                    query: _query,
                   }
                 : data.searchResult,
             ...props,
@@ -168,7 +168,7 @@ const SearchContext = props => {
 
     return React.cloneElement(props.children, {
       searchResult: {
-        query: props.params.query,
+        query: _query,
       },
       ...props,
       ...vtexSearchResult,