+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Change querystring from `query` to `_query`
+
 ## [0.2.0] - 2019-11-14
 
 ### Added